moved _get_module_with_only_relevant_foreach_idx to module_handler as this is the only place where it's used
