+++ tests/terraform/graph/variable_rendering/test_foreach_renderer.py
@@ -433,7 +433,7 @@ def test__get_module_with_only_relevant_foreach_idx():
                       )
     original_key = TFModule(name='2', path='2', foreach_idx='2',
                             nested_tf_module=TFModule(name='3', path='3', foreach_idx='3', nested_tf_module=None))
-    result = ForeachAbstractHandler._get_module_with_only_relevant_foreach_idx('test', original_key, module)
+    result = ForeachModuleHandler._get_module_with_only_relevant_foreach_idx('test', original_key, module)
     assert result == TFModule(name='1', path='1', foreach_idx='1',
                       nested_tf_module=TFModule(name='2', path='2', foreach_idx='test',
                                                 nested_tf_module=TFModule(name='3', path='3', foreach_idx='3',
